+++ javascript/example/index.js
@@ -188,7 +188,7 @@ document.addEventListener('WebComponentsReady', () => {
 
     const modelLoader = new THREE.ModelLoader(viewer.loadingManager);
     viewer.urdfLoader.defaultMeshLoader = (path, ext, done) => {
-        modelLoader.load(path, res => done(res.model));
+        modelLoader.load(path, res => done(res.model), null, err => done(null, err));
     };
 
     document.querySelector('li[urdf]').dispatchEvent(new Event('click'));