+++ javascript/src/URDFLoader.js
@@ -128,12 +128,13 @@ class URDFLoader {
 
         let result = null;
         let meshCount = 0;
-        const loadMeshFunc = (path, ext, done) => {
 
-            meshCount++;
-            options.loadMeshCb(path, ext, (...args) => {
+        const createMeshTallyFunc = func => {
 
-                done(...args);
+            return (...args) => {
+
+                func(...args);
+
                 meshCount--;
                 if (meshCount === 0) {
 
@@ -144,8 +145,13 @@ class URDFLoader {
                     });
 
                 }
+            };
+        };
 
-            });
+        const loadMeshFunc = (path, ext, done) => {
+
+            meshCount++;
+            options.loadMeshCb(path, ext, createMeshTallyFunc(done));
 
         };
         result = this._processUrdf(content, packages, options.workingPath, loadMeshFunc);
@@ -467,9 +473,13 @@ class URDFLoader {
                         const scaleAttr = n.children[0].getAttribute('scale');
                         if (scaleAttr) scale = this._processTuple(scaleAttr);
 
-                        loadMeshCb(filePath, ext, obj => {
+                        loadMeshCb(filePath, ext, (obj, err) => {
 
-                            if (obj) {
+                            if (err) {
+
+                                console.error('URDFLoader: Error loading mesh.', err);
+
+                            } else if (obj) {
 
                                 if (obj instanceof THREE.Mesh) {